requirements.txt
slack-bolt>=1.14.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
certifi>=2021.0.0
httplib2>=0.20.0
